langgrap_agentic/customer_support/requirements.txt
langgraph
langsmith
langchain_community
langchain_openai
tavily-python
pandas
jupyterlab
python-dotenv